ISSUE-36: almost all kafka entities in package hu.exercise.spring.kafka.topic
