+++ src/main/java/hu/exercise/spring/kafka/cogroup/Source.java
@@ -1,0 +1,5 @@
+package hu.exercise.spring.kafka.cogroup;
+
+public enum Source {
+	DB, TSV;
+}
