--- src/main/java/hu/exercise/spring/kafka/event/Source.java
@@ -1,5 +1,0 @@
-package hu.exercise.spring.kafka.event;
-
-public enum Source {
-	DB, TSV;
-}
