+++ src/main/java/hu/exercise/spring/kafka/listener/FlushedMessageListener.java
@@ -1,7 +1,5 @@
 package hu.exercise.spring.kafka.listener;
 
-import java.util.concurrent.atomic.AtomicLong;
-
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.DisposableBean;
@@ -11,8 +9,8 @@ import org.springframework.stereotype.Service;
 
 import hu.exercise.spring.kafka.KafkaEnvironment;
 import hu.exercise.spring.kafka.ShutdownController;
-import hu.exercise.spring.kafka.cogroup.Flushed;
 import hu.exercise.spring.kafka.cogroup.Report;
+import hu.exercise.spring.kafka.topic.Flushed;
 import io.github.springwolf.core.asyncapi.annotations.AsyncListener;
 import io.github.springwolf.core.asyncapi.annotations.AsyncOperation;
 import io.reactivex.rxjava3.annotations.NonNull;
