+++ src/main/java/hu/exercise/spring/kafka/topic/Flushed.java
@@ -1,25 +1,33 @@
-package hu.exercise.spring.kafka.event;
+package hu.exercise.spring.kafka.topic;
 
-import hu.exercise.spring.kafka.cogroup.Action;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
+import lombok.NoArgsConstructor;
 import net.csini.spring.kafka.KafkaEntity;
 import net.csini.spring.kafka.KafkaEntityKey;
 
 @Data
+@Builder
+@NoArgsConstructor
 @AllArgsConstructor
-@EqualsAndHashCode
+@EqualsAndHashCode(of = "requestid")
 @KafkaEntity
-public class DBEvent {
+public class Flushed {
 
 	@KafkaEntityKey
 	@Schema(description = "Run's unique identifier", example = "a3dbaa5a-1375-491e-8c21-403864de8779")
 	private String requestid;
-	
-	@Schema(description = "Your product’s unique identifier", example = "A2B4")
-	private String id;
-	
-	private Action action;
+
+	private int countInsert;
+
+	private int countUpdate;
+
+	private int countDelete;
+
+	private int countError;
+
+	private int sumProcessed;
 }
