+++ src/main/java/hu/exercise/spring/kafka/topic/ReadedFromDBEvent.java
@@ -1,8 +1,8 @@
-package hu.exercise.spring.kafka.cogroup;
+package hu.exercise.spring.kafka.topic;
 
+import hu.exercise.spring.kafka.input.Product;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AllArgsConstructor;
-import lombok.Builder;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
@@ -10,24 +10,19 @@ import net.csini.spring.kafka.KafkaEntity;
 import net.csini.spring.kafka.KafkaEntityKey;
 
 @Data
-@Builder
-@NoArgsConstructor
 @AllArgsConstructor
-@EqualsAndHashCode(of = "requestid")
+@NoArgsConstructor
+@EqualsAndHashCode(of = { "id", "requestid" })
 @KafkaEntity
-public class Flushed {
-
+public class ReadedFromDBEvent {
+	
+	@Schema(description = "Your product’s unique identifier", example = "A2B4")
+	private String id;
+	
 	@KafkaEntityKey
 	@Schema(description = "Run's unique identifier", example = "a3dbaa5a-1375-491e-8c21-403864de8779")
 	private String requestid;
+	
+	private Product product;
 
-	private int countInsert;
-
-	private int countUpdate;
-
-	private int countDelete;
-
-	private int countError;
-
-	private int sumProcessed;
 }
